requirements.txt
llama-index-core
llama-index-vector-stores-chroma
llama-index-embeddings-ollama
llama-index-llms-ollama
chromadb
ollama